Linear Regression model to predict housing prices based on traning dataset
A simple linear regression model that would use the training dataset to train itself and predict house prices based on variables like number of bathrooms and bedrooms and total square feet.

Currently the model is still under development as it predicts the prices but the accuracy is not that great. Rest assured I'll try to keep updating it time to time and get to the best accuracy possible.